修复Integer封装类做比较的bug

+++ src/com/fly/controller/beauty/BeautyController.java
@@ -234,7 +234,7 @@ public class BeautyController extends BaseController
             return;
         }
         User user = Constants.getLoginUser(getSession());
-        if (user.getInt("id") != beauty.getInt("user_id"))
+        if (user.getInt("id").intValue() != beauty.getInt("user_id").intValue())
         {
             message = "非法操作";
             setAttr("message", message);
